final changes for consistency
finally fixed either the final batch of unused styles or inconsistent formatting. Simplified last step of the header menu to make sure it as readable as possible. made sure to put ids for every unique item and get rid of the classes that were there previous to make sure to not mess up styling later. it is now a safe state to pull from for the feature branches.