yall can start working now!!!
YAYYY we finally did it were so good at coding! this makes so much sense and im so proud to be going into college next year as a comp sci major/ software engineer!! this is so exciting and makes me feel so good about myself and my abillities. we did so well today and I am so proud of all of us. Please save me.