fsefsfes
I got a small bug fix done and that took me a good bit to realize why something was being weird on the pages. but I ended up having to restart my progress on the components because I couldnt figure out what was up with it. I should be good to lock in with it tomorrow.